Replace with Singleton<T> calling methods
Regex replaced the dependencies to use Singleton<T> calling instead.

\[Imported\] (.*?(?=\s))(\s*!?(?=\s))\s(.*?(?= ))(\s*!?(?=\s))\s(.*?(?=;));
readonly $1 $3 $5 = Singleton<$3>();

A few had to be made non readonly.
Extra sacrifices had to be made too.

The idea is we sacrifice in terms of having a "static god dependency resolver" in favor of allowing constructors to work again normally. Runtime errors due to constructors failing is more annoying otherwise, and likely  New<T> was incredibly inefficient due to the field reflection.